RR 2.6.8.1-EXP XML Update
Better Dungeons 2 is not yet included despite the integration of its configs as a base to work with.

Instead this is a general update: Better Dungeons 2 integration will require a dedicated update.

Also, after testing, the promised output corrections to AtSci have gone through in this build, thus the conversion ratio has been reset to the defaults.