Config update pass part 1 for 2.6.10.0-EXP
I ran into trouble with RI so the update has been delayed until I either get a fix or decide to pull the mod.

RI is disabled temporarily in the meantime to do maintenance to other mods.

The IC2 suite is being removed in this update after careful consideration of player feedback.