Repairing goofed Mariculture IDs for future 2.5 player migration
Well this was a dumb and easily preventable oversight. It happens I guess.

The IDs have been moved back to their settings in 2.5.6.1. New IDs are left alone - I presume this happened because the config changes were so extreme it masked the stuff that still existed.

To combat this, and also because every active world now needs it, retrogen for Mariculture is now activated. This will be de-activated again before release as 2.5.6.1 worlds should migrate cleanly without damage after this change.