BD2 spawn rate changes
The consensus is that BD2 is too "OP", although people have trouble placing exactly why and it's subjective, like everything.

What did become clear is that the spawn rate of the dungeons is currently too frequent... or rather it's either too rare or way too frequent, averaging out to merely "OK".

This set of changes aims to fix the weird clustering behaviour we're seeing. If in the process dungeons become too rare we will accomodate that accordingly.

Additionally, as a bug fix attempt, the loot added directly by BD2 is disabled (most of it being grossly underpowered in the pack anyway).