UE 3.1 incompatibility rollback hotfix
A couple of our mods have not updated to UE 3.1 yet, and any attempt to load a world with these UE 3.0 mods caused painful crash issues.

The entire UE suite has been rolled back to the last UE 3.0 builds until everyone gets with the program, as the 3.1 bugfix is actually quite significant and worth the upgrade.