RR 2.6.16.0-EXP XML update
The mod removals in this release bumped it up from a minor release change.

Removed Agriculture as it proved to be a third wheel in the pack after GrowthCraft and Magical Crops tag teaming to handle farming so well.

Removed BackTools as it was not working out anywhere near as well as expected.