Evaluation completion for 2.7.0.0-RR release
(Note that Advanced Genetics was evaluated internally by Dave and will be added late as part of this update.)

Dynamic Liquid Tanks 2 and Schematica have failed. Everything else has passed.

GrowthCraft's Fishtrap submodule has been removed in order to resolve the recipe conflict and unblock the major fishing submodule of Mariculture (fish breeding!)

Advanced Thaumaturgy passed but this pass is contingent on it staying in-date which it is currently failing to do. When it updates to TC4.1 it will be reinstated at that point.