Disabling the crafting recipe for the Mariculture Vat
There is a recipe conflict between this item and the Copper Canister from Galacticraft. Copper Canisters are used to produce Refineries.

After consulting with the staff channel, it was agreed that both blocks are vital to their mods, so one of the recipes has to be moved.

This is part one of that process.