Initial skeleton commit
This gets the fundamentals of the pack confirmed operational so we have a foundation to build upon. No attempt at tuning the mods has been made.

Note that old files will be reinstated if they remain relevant to the 3.0 branch.

Generally though it's going to be far more effective to start fresh.