2-DEV-160 XML update part 2
Note that Truss Mod has been removed (this got accidentally rolled into part 1).

We didn't manage to get perms for AnimationAPI (in fact it looks like the dev is MIA!) so we have to resort to the workaround for now.

Mods that passed evaluation have been relocated.

Will be used to cut stable releases as 2.9.2.2